requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
debugpy==1.8.14
Flask==2.3.3
Flask-HTTPAuth==4.8.0
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.24.0
google-cloud-storage==2.16.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.72.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
